Update allow_file_group_writeable with more flexibility
Case EA-4868: This updates the allow_file_group_writeable config
option with 3 values, instead of the boolean true/false.  This can
now also be the value 'cpanel'.

Traditionally, this was a simple test case.  Though by default,
cPanel configures suphp to the same user/group of the owner
of the domain.  Thus, we could turn off allow_file_group_writeable
and it really doesn't hurt anything security-wise.

Unfortunately, it does prevent Windows users who copy files via
WinSFTP to break since the permissions it defaults to are 0666;
and after the applied umask it's 0664.  Either way, suphp won't
run it because the group write bit is set.

This patch updates the setting with a 'cpanel' option to go
ahead and allow the group write bit to be set, but only if the
user is the only person in the group (specifically, no other
users defined in the user's group within /etc/group).

If you assign multiple users to the same group in /etc/passwd,
this then check is useless and this option doesn't apply
to you.